backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aistudio-sdk==0.3.8
alembic==1.18.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bce-python-sdk==0.9.59
bcrypt==5.0.0
beautifulsoup4==4.14.3
cachetools==6.2.4
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
cssselect==1.3.0
cssutils==2.11.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
einops==0.8.1
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.128.0
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.1.0
ftfy==6.3.1
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.187.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-genai==1.60.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.3.2
idna==3.11
imagesize==1.4.1
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.83
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langsmith==0.6.4
loguru==0.7.3
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==3.26.2
modelscope==1.34.0
more-itertools==10.8.0
multidict==6.7.0
mypy_extensions==1.1.0
# networkx==3.6.1
# numpy==2.4.1
# opencv-contrib-python==4.10.0.84
opencv-python-headless==4.13.0.90
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.5
packaging==25.0
# paddleocr==3.3.3
# paddlepaddle==3.2.2
# paddlex==3.3.13
paddleocr==3.1.0
paddlepaddle==3.1.0
paddlex==3.1.0
pandas==2.3.3
passlib==1.7.4
pillow==12.1.0
premailer==3.10.0

pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyMuPDF==1.26.3
pyparsing==3.3.2
pypdfium2==5.3.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
ruamel.yaml==0.19.1
safetensors==0.7.0
# scikit-learn==1.8.0
# scipy==1.17.0
# sentencepiece==0.2.1
# setuptools==80.9.0
# shapely==2.1.2
# shellingham==1.5.4
# six==1.17.0
# sniffio==1.3.1
# soupsieve==2.8.3
SQLAlchemy==2.0.45
starlette==0.50.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.1
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
wcwidth==0.2.14
websockets==15.0.1
yarl==1.22.0
zstandard==0.25.0
redis==6.4.0
celery==5.6.2

psycopg2==2.9.11